Main is now a common initialisation point and most initialisation is now in ServerMain
