+++ src/main/java/de/fisch37/betterserverpacksfabric/PackCommand.java
@@ -33,7 +33,7 @@ public class PackCommand {
 
     private static LiteralArgumentBuilder<ServerCommandSource> makeCommand(CommandRegistryAccess registryAccess) {
         return literal("pack")
-                .requires(Main::hasConfigAccess)
+                .requires(ServerMain::hasConfigAccess)
                 .then(literal("set")
                         .executes(PackCommand::disablePack)
                         .then(argument("url", string())
@@ -88,7 +88,7 @@ public class PackCommand {
                 ,
                 true
         );
-        Main.updateHash().thenAccept(result -> {
+        ServerMain.updateHash().thenAccept(result -> {
             // Let's all hope that this doesn't cause threading issues :+1:
             if (result == null) {
                 // Error
@@ -150,7 +150,7 @@ public class PackCommand {
             return 0;
         }
 
-        Main.config.url.set(url).save();
+        ServerMain.config.url.set(url).save();
         sendUpdate(context);
         source.sendFeedback( () -> MSG_PREFIX.copy()
                 .append("Pack URL has been updated. Reloading hash...")
@@ -162,7 +162,7 @@ public class PackCommand {
     }
 
     private static int disablePack(CommandContext<ServerCommandSource> context) {
-        Main.config.url.set("").save();
+        ServerMain.config.url.set("").save();
         updateHashWithContext(context.getSource(), false);
         return 1;
     }
@@ -175,7 +175,7 @@ public class PackCommand {
     private static int getRequired(CommandContext<ServerCommandSource> context) {
         ServerCommandSource source = context.getSource();
 
-        Boolean required = Main.config.required.get();
+        Boolean required = ServerMain.config.required.get();
 
         source.sendFeedback(() -> MSG_PREFIX.copy()
                 .append("Pack is " + (required ? "required" : "optional"))
@@ -189,7 +189,7 @@ public class PackCommand {
         ServerCommandSource source = context.getSource();
 
         Boolean required = BoolArgumentType.getBool(context, "required");
-        Main.config.required.set(required).save();
+        ServerMain.config.required.set(required).save();
         sendUpdate(context);
 
         source.sendFeedback( () -> MSG_PREFIX.copy()
@@ -201,7 +201,7 @@ public class PackCommand {
     }
 
     private static int showPrompt(CommandContext<ServerCommandSource> context) {
-        final Optional<Text> prompt = Main.config.getPrompt(context.getSource().getRegistryManager());
+        final Optional<Text> prompt = ServerMain.config.getPrompt(context.getSource().getRegistryManager());
         context.getSource().sendFeedback(
                 () -> prompt.map(
                         text -> MSG_PREFIX.copy()
@@ -218,7 +218,7 @@ public class PackCommand {
 
     private static int setPrompt(CommandContext<ServerCommandSource> context) {
         final Text prompt = TextArgumentType.getTextArgument(context, "prompt");
-        Main.config.setPrompt(prompt, context.getSource().getRegistryManager())
+        ServerMain.config.setPrompt(prompt, context.getSource().getRegistryManager())
                 .save();
         sendUpdate(context);
         context.getSource().sendFeedback(
@@ -232,7 +232,7 @@ public class PackCommand {
     }
 
     private static int clearPrompt(CommandContext<ServerCommandSource> context) {
-        Main.config.setPrompt(null, null);
+        ServerMain.config.setPrompt(null, null);
         sendUpdate(context);
         context.getSource().sendFeedback(
                 () -> MSG_PREFIX.copy()
@@ -244,9 +244,9 @@ public class PackCommand {
     }
 
     private static int showInfo(CommandContext<ServerCommandSource> context) {
-        String url = Main.config.url.get();
-        boolean required = Main.config.required.get();
-        Optional<Text> prompt = Main.config.getPrompt(context.getSource().getRegistryManager());
+        String url = ServerMain.config.url.get();
+        boolean required = ServerMain.config.required.get();
+        Optional<Text> prompt = ServerMain.config.getPrompt(context.getSource().getRegistryManager());
         if (!url.isEmpty()) {
             context.getSource().sendFeedback(() ->
                             MSG_PREFIX.copy()
@@ -260,7 +260,7 @@ public class PackCommand {
                                     )
                                     .append("\n")
                                     .append("Pack hash: ")
-                                    .append(Optional.ofNullable(Main.getHashString())
+                                    .append(Optional.ofNullable(ServerMain.getHashString())
                                             .map(s -> Text.literal(s)
                                                     .formatted(Formatting.LIGHT_PURPLE)
                                                     .formatted(Formatting.ITALIC)
