+++ src/main/java/de/fisch37/betterserverpacksfabric/client/MainClient.java
@@ -9,7 +9,6 @@ import io.wispforest.owo.util.Observable;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
-import net.minecraft.client.toast.SystemToast;
 import net.minecraft.text.Text;
 import org.jetbrains.annotations.Nullable;
