+++ src/main/java/de/fisch37/betterserverpacksfabric/networking/Networking.java
@@ -1,16 +1,12 @@
 package de.fisch37.betterserverpacksfabric.networking;
 
-import de.fisch37.betterserverpacksfabric.Main;
+import de.fisch37.betterserverpacksfabric.ServerMain;
 import io.wispforest.owo.network.OwoNetChannel;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 
-import java.util.Collection;
-import java.util.stream.Collectors;
-
 import static de.fisch37.betterserverpacksfabric.Main.MOD_ID;
-import static de.fisch37.betterserverpacksfabric.Main.config;
+import static de.fisch37.betterserverpacksfabric.ServerMain.config;
 
 public class Networking {
     public static final OwoNetChannel CHANNEL = OwoNetChannel.create(Identifier.of(MOD_ID, "main"));
@@ -19,7 +15,7 @@ public class Networking {
         CHANNEL.registerClientboundDeferred(CanChangeConfigPacket.class);
         CHANNEL.registerClientboundDeferred(PackState.class);
         CHANNEL.registerServerbound(PackState.class, (packet, access) -> {
-            if (Main.hasConfigAccess(access.player())) {
+            if (ServerMain.hasConfigAccess(access.player())) {
                 config.url.set(packet.url()).save();
                 config.required.set(packet.required()).save();
                 config.rehashOnStart.set(packet.rehashNextStart()).save();
@@ -36,7 +32,7 @@ public class Networking {
         var targets = server.getPlayerManager()
                 .getPlayerList()
                 .stream()
-                .filter(Main::hasConfigAccess)
+                .filter(ServerMain::hasConfigAccess)
                 .toList();
         CHANNEL.serverHandle(targets).send(PackState.fromConfig(config, server.getRegistryManager()));
     }
